scan: Extract scanners to a support module
